+++ gbdxtools/catalog.py
@@ -66,43 +66,49 @@ class Catalog():
         lng = results[0]['geometry']['location']['lng']
         return lat, lng
 
-    def search_address(self, address, type="Acquisition"):
+    def search_address(self, address, filters=None, startDate=None, endDate=None, types=None):
         ''' Perform a catalog search over an address string
 
         Args:
             address: any address string
-            type: type, default=Acquisition
+            filters: Array of filters.  Optional.  Example:
+                [  
+                    "(sensorPlatformName = 'WORLDVIEW01' OR sensorPlatformName ='QUICKBIRD02')",
+                    "cloudCover < 10",
+                    "offNadirAngle < 10"
+                ]
+            startDate: string.  Optional.  Example: "2004-01-01T00:00:00.000Z"
+            endDate: string.  Optional.  Example: "2004-01-01T00:00:00.000Z"
+            types: Array of types to search for.  Optional.  Example (and default):  ["Acquisition"]
 
         Returns:
-            catalog resultset dictionary
+            catalog search resultset
         '''
         lat, lng = self.get_address_coords(address)
-        return self.search_point(lat,lng, type)
+        return self.search_point(lat,lng, filters=filters, startDate=startDate, endDate=endDate, types=types)
 
-    def search_point(self, lat, lng, type="Acquisition"):
+    #def search_point(self, lat, lng, type="Acquisition")
+    def search_point(self, lat, lng, filters=None, startDate=None, endDate=None, types=None, type=None):
         ''' Perform a catalog search over a specific point, specified by lat,lng
 
         Args:
             lat: latitude
             lng: longitude
-            type: type, default=Acquisition
+            filters: Array of filters.  Optional.  Example:
+                [  
+                    "(sensorPlatformName = 'WORLDVIEW01' OR sensorPlatformName ='QUICKBIRD02')",
+                    "cloudCover < 10",
+                    "offNadirAngle < 10"
+                ]
+            startDate: string.  Optional.  Example: "2004-01-01T00:00:00.000Z"
+            endDate: string.  Optional.  Example: "2004-01-01T00:00:00.000Z"
+            types: Array of types to search for.  Optional.  Example (and default):  ["Acquisition"]
 
         Returns:
-            catalog resultset dictionary
+            catalog search resultset
         '''
-        postdata = {  
-            "searchAreaWkt":"POLYGON ((%s %s, %s %s, %s %s, %s %s, %s %s))" % (lng, lat,lng,lat,lng,lat,lng,lat,lng,lat),
-            "filters":[  
-            ],
-            "types":[  
-                type
-            ]
-        }
-        url = 'https://geobigdata.io/catalog/v1/search?includeRelationships=false'
-        headers = {'Content-Type':'application/json'}
-        r = self.gbdx_connection.post(url, headers=headers, data=json.dumps(postdata))
-        r.raise_for_status()
-        return r.json()
+        searchAreaWkt = "POLYGON ((%s %s, %s %s, %s %s, %s %s, %s %s))" % (lng, lat,lng,lat,lng,lat,lng,lat,lng,lat)
+        return self.search(searchAreaWkt=searchAreaWkt, filters=filters, startDate=startDate, endDate=endDate, types=types)
 
     def get_data_location(self, catalog_id):
         """
@@ -257,10 +263,8 @@ class Catalog():
         Returns:
             single catalog item, or none if not found
         '''
-        if not len(results['results']):
+        if not len(results):
             return None
-
-        results = results['results']
 
         # filter on type
         if types: